# gpt/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pathfinder 1E Master DD Core API",
        "version": "1.0.0",
        "description": "API minimale per esporre i moduli e le risorse del kernel Master DD a un GPT tramite Actions."
    },
    "servers": [
        {
            "url": "https://YOUR-SERVER-URL",
            "description": "Sostituisci con l'URL reale del tuo server FastAPI"
        }
    ],
    "paths": {
...
        "/modules": {
            "get": {
                "operationId": "listModules",
                "summary": "Lista dei file modulo disponibili",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista di moduli",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "size_bytes": {
                                                "type": "integer"
                                            },
                                            "suffix": {
                                                "type": "string"
                                            }
                                        },
                                        "required": [
                                            "name"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/build/stub": {
            "post": {
                "operationId": "buildStub",
                "summary": "Genera uno stub deterministico per il MinMax Builder",
                "description": "Restituisce un payload JSON stub per la costruzione di un personaggio. Equivalente allo stub precedentemente richiesto tramite /modules/minmax_builder.txt?mode=stub.",
                "parameters": [
                    {
                        "name": "mode",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "extended"
                        },
                        "description": "Modalita' builder (core/extended)"
                    },
                    {
                        "name": "class",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Classe del personaggio"
                    },
                    {
                        "name": "race",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Razza del personaggio"
                    },
                    {
                        "name": "archetype",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Archetipo del personaggio"
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Livello del personaggio"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payload JSON stub",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Errore interno durante la generazione dello stub"
                    }
                }
            }
        },
        "/modules/{name}": {
            "get": {
                "operationId": "getModuleContent",
                "summary": "Contenuto testuale di un modulo",
                "description": "Restituisce il testo del modulo. Se ALLOW_MODULE_DUMP=false lato server, il contenuto viene troncato a 4000 caratteri con un marcatore finale. Supporta anche stub per il minmax builder.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Nome del file modulo (es. base_profile.txt)"
                    },
                    {
                        "name": "mode",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "extended"
                        },
                        "description": "Modalita' builder (core/extended/full-pg)"
                    },
                    {
                        "name": "class",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Classe per lo stub minmax builder"
                    },
                    {
                        "name": "race",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Razza per lo stub minmax builder"
                    },
                    {
                        "name": "archetype",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Archetipo per lo stub minmax builder"
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Livello per lo stub minmax builder"
                    },
                    {
                        "name": "stub",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        },
                        "description": "Restituisce lo stub payload per il minmax builder"
                    }
                ],
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contenuto testuale o JSON stub",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Modulo non trovato"
                    }
                }
            },
            "post": {
                "operationId": "postModuleContent",
                "summary": "Contenuto testuale di un modulo (modalita' POST)",
                "description": "Equivalente POST di /modules/{name}; utile per inviare body JSON per lo stub builder.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Nome del file modulo"
                    },
                    {
                        "name": "mode",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "extended"
                        }
                    },
                    {
                        "name": "class",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "race",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "archetype",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stub",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contenuto testuale o JSON stub",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Modulo non trovato"
                    }
                }
            }
        },
        "/modules/{name}/meta": {
            "get": {
                "operationId": "getModuleMeta",
                "summary": "Metadati per un modulo",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadati del modulo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "size_bytes": {
                                            "type": "integer"
                                        },
                                        "suffix": {
                                            "type": "string"
                                        },
                                        "version": {
                                            "type": "string"
                                        },
                                        "compatibility": {
                                            "oneOf": [
                                                {"type": "string"},
                                                {"type": "object"},
                                                {"type": "array"}
                                            ]
                                        }
                                    },
                                    "required": [
                                        "name"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Modulo non trovato"
                    }
                }
            }
        },
        "/modules/taverna_saves/meta": {
            "get": {
                "operationId": "getTavernaSavesMeta",
                "summary": "Metadati directory taverna_saves",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadati e policy di storage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/modules/taverna_saves/quota": {
            "get": {
                "operationId": "getTavernaSavesQuota",
                "summary": "Quota e utilizzo taverna_saves",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metriche quota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/taverna_storage_meta": {
            "get": {
                "operationId": "getTavernaStorageMeta",
                "summary": "Metadati storage taverna_saves",
                "description": "Restituisce metadati, quota e policy della directory taverna_saves.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadati storage taverna_saves",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/storage_meta": {
            "get": {
                "operationId": "getStorageMeta",
                "summary": "Alias legacy per /taverna_storage_meta",
                "description": "Mantiene la compatibilità con le azioni GPT esistenti.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadati storage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key",
                "description": "Chiave API condivisa impostata tramite variabile d'ambiente API_KEY; richiesta salvo ALLOW_ANONYMOUS=true"
            }
        }
    }
}